change main wallet layout
