+++ components/AccountTabs.tsx
@@ -6,9 +6,9 @@ import BalanceCard from './BalanceCard';
 import Text from './Text';
 import React, {useCallback, useState} from 'react';
 import {BalanceFragment, Connection_Stats_Enum} from '../constants/Type';
-import {useBottomSheet} from '../hooks/useBottomSheet';
+import {useBottomSheet} from '../src/hooks/useBottomSheet';
 import BottomSheetMenu from './BottomSheetMenu';
-import useWallet from '../hooks/useWallet';
+import useWallet from '../src/hooks/useWallet';
 import {
   StyleService,
   useStyleSheet,
@@ -43,9 +43,27 @@ const AccountsTab = () => {
             backgroundColor: theme['color-primary-100'],
           },
         }}>
-        <TabNavigator.Screen name="Accounts" component={Accounts} listeners={({navigation}) => ({blur: () => navigation.setParams({screen: undefined})})}/>
-        <TabNavigator.Screen name="Tokens" component={Tokens} listeners={({navigation}) => ({blur: () => navigation.setParams({screen: undefined})})}/>
-        <TabNavigator.Screen name="NFTs" component={NFTs} listeners={({navigation}) => ({blur: () => navigation.setParams({screen: undefined})})}/>
+        <TabNavigator.Screen
+          name="Accounts"
+          component={Accounts}
+          listeners={({navigation}) => ({
+            blur: () => navigation.setParams({screen: undefined}),
+          })}
+        />
+        <TabNavigator.Screen
+          name="Tokens"
+          component={Tokens}
+          listeners={({navigation}) => ({
+            blur: () => navigation.setParams({screen: undefined}),
+          })}
+        />
+        <TabNavigator.Screen
+          name="NFTs"
+          component={NFTs}
+          listeners={({navigation}) => ({
+            blur: () => navigation.setParams({screen: undefined}),
+          })}
+        />
       </TabNavigator.Navigator>
     </View>
   );
@@ -159,9 +177,14 @@ const Accounts = () => {
       </>,
     );
   }, [accounts]);
-  
+
   return (
-    <View style={{backgroundColor: theme['color-basic-700'], flex: 1}}>
+    <View
+      style={{
+        backgroundColor: theme['color-basic-700'],
+        flex: 1,
+        marginTop: scale(-6),
+      }}>
       <ScrollView
         refreshControl={
           <RefreshControl
@@ -278,7 +301,12 @@ const NFTs = () => {
   }, [nfts]);
 
   return (
-    <View style={{backgroundColor: theme['color-basic-700'], flex: 1}}>
+    <View
+      style={{
+        backgroundColor: theme['color-basic-700'],
+        flex: 1,
+        marginTop: scale(-6),
+      }}>
       <ScrollView
         refreshControl={
           <RefreshControl
@@ -377,7 +405,12 @@ const Tokens = () => {
   }, [tokens]);
 
   return (
-    <View style={{backgroundColor: theme['color-basic-700'], flex: 1}}>
+    <View
+      style={{
+        backgroundColor: theme['color-basic-700'],
+        flex: 1,
+        marginTop: scale(-6),
+      }}>
       <ScrollView
         refreshControl={
           <RefreshControl
