+++ components/BottomSheetMenu.tsx
@@ -6,7 +6,7 @@ import {StyleService, useStyleSheet} from '@tsejerome/ui-kitten-components';
 import {NavigationProp, useNavigation} from '@react-navigation/native';
 import {RootStackParamList, WalletParamList} from '../navigation/type';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
-import {useBottomSheet} from '../hooks/useBottomSheet';
+import {useBottomSheet} from '../src/hooks/useBottomSheet';
 
 const BottomSheetMenu = (props: any) => {
   const styles = useStyleSheet(themedStyles);
