+++ components/BottomSheetOptions.tsx
@@ -4,7 +4,7 @@ import {View} from 'react-native';
 import React, {memo, useMemo, useState} from 'react';
 import {StyleService, useStyleSheet} from '@tsejerome/ui-kitten-components';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
-import {useBottomSheet} from '../hooks/useBottomSheet';
+import {useBottomSheet} from '../src/hooks/useBottomSheet';
 import {scale, verticalScale} from 'react-native-size-matters';
 
 const BottomSheetOptions = (props: any) => {
