+++ components/ScanQR.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import {TouchableOpacity, View} from 'react-native';
 import QRCodeScanner from 'react-native-qrcode-scanner';
 import {scale} from 'react-native-size-matters';
-import {useBottomSheet} from '../hooks/useBottomSheet';
+import {useBottomSheet} from '../src/hooks/useBottomSheet';
 import BottomSheetView from './BottomSheetView';
 import OptionCard from './OptionCard';
 import Text from './Text';
