+++ contexts/SecurityProvider.tsx
@@ -15,12 +15,12 @@ import {Text} from '@tsejerome/ui-kitten-components';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 import Container from '../components/Container';
 import {scale} from 'react-native-size-matters';
-import useWallet from '../hooks/useWallet';
+import useWallet from '../src/hooks/useWallet';
 import {AppState, Platform} from 'react-native';
 import {useNavigation} from '@react-navigation/native';
 import {check, PERMISSIONS, RESULTS, request} from 'react-native-permissions';
 import {Connection_Stats_Enum} from '../constants/Type';
-import useTraceUpdates from '../hooks/useTraceUpdates';
+import useTraceUpdates from '../src/hooks/useTraceUpdates';
 
 export const SecurityProvider = (props: any) => {
   const [lockedScreen, setLockedScreen] = useState(false);
@@ -88,14 +88,7 @@ export const SecurityProvider = (props: any) => {
           nextAppState === 'active' &&
           appState.current === 'background')
       ) {
-        if (
-          refreshWallet &&
-          !(
-            connected == Connection_Stats_Enum.Bootstrapping ||
-            connected == Connection_Stats_Enum.Connecting ||
-            connected == Connection_Stats_Enum.Syncing
-          )
-        ) {
+        if (refreshWallet) {
           refreshWallet();
         }
         if (lockAfterBackground === true) {
