+++ src/hooks/useQr.tsx
@@ -1,5 +1,5 @@
 import {useContext} from 'react';
-import {QrContext} from '../contexts/QrContext';
+import {QrContext} from '../../contexts/QrContext';
 
 export const useQr = () => {
   const context = useContext(QrContext);
