+++ src/screens/wallet/AddStakingNodeScreen.tsx
@@ -7,16 +7,17 @@ import {
   Layout,
   TopNavigation,
 } from '@tsejerome/ui-kitten-components';
-import Container from '../../components/Container';
+import Container from '../../../components/Container';
 import {NavigationProp, useNavigation} from '@react-navigation/native';
-import {NodeOption} from '../../constants/Type';
-import Text from '../../components/Text';
-import {RootStackParamList} from '../../navigation/type';
+import {NodeOption} from '../../../constants/Type';
+import Text from '../../../components/Text';
+import {RootStackParamList} from '../../../navigation/type';
 import {KeyboardAwareScrollView} from 'react-native-keyboard-aware-scroll-view';
-import TopNavigationComponent from '../../components/TopNavigation';
+import TopNavigationComponent from '../../../components/TopNavigation';
 
 const AddStakingNodeScreen = () => {
-  const {updateAccounts, ExecWrapperPromise, ExecWrapperSyncPromise} = useWallet();
+  const {updateAccounts, ExecWrapperPromise, ExecWrapperSyncPromise} =
+    useWallet();
 
   const {goBack} = useNavigation<NavigationProp<RootStackParamList>>();
 
@@ -52,7 +53,9 @@ const AddStakingNodeScreen = () => {
     }
     if (
       !(await ExecWrapperSyncPromise(
-        'bitcore.Address("'+newNode.address+'").isPayToPublicKeyHash'))) {
+        'bitcore.Address("' + newNode.address + '").isPayToPublicKeyHash',
+      ))
+    ) {
       setError('You need to specify a NAV address from the staking node');
       return;
     }
