+++ src/screens/wallet/MintNftScreen.tsx
@@ -1,35 +1,37 @@
 import useWallet from '../../hooks/useWallet';
 import React, {useEffect, useState} from 'react';
-import {View, StyleSheet} from 'react-native';
+import {StyleSheet, View} from 'react-native';
 import {
   Button,
   Input,
   Layout,
   TopNavigation,
-  useStyleSheet,
 } from '@tsejerome/ui-kitten-components';
-import Container from '../../components/Container';
+import Container from '../../../components/Container';
 import {NavigationProp, useNavigation} from '@react-navigation/native';
-import {CollectionOption} from '../../constants/Type';
-import Text from '../../components/Text';
-import {RootStackParamList} from '../../navigation/type';
+import {Destination_Types_Enum, NftItemOption} from '../../../constants/Type';
+import Text from '../../../components/Text';
+import {RootStackParamList} from '../../../navigation/type';
 import {KeyboardAwareScrollView} from 'react-native-keyboard-aware-scroll-view';
-import TopNavigationComponent from '../../components/TopNavigation';
+import TopNavigationComponent from '../../../components/TopNavigation';
 import useSecurity from '../../hooks/useSecurity';
-import BottomSheetView from '../../components/BottomSheetView';
-import {SwipeButton} from '../../components/SwipeButton';
+import BottomSheetView from '../../../components/BottomSheetView';
+import {SwipeButton} from '../../../components/SwipeButton';
 import {useBottomSheet} from '../../hooks/useBottomSheet';
 import {useModal} from '../../hooks/useModal';
-import LoadingModalContent from '../../components/Modals/LoadingModalContent';
+import LoadingModalContent from '../../../components/Modals/LoadingModalContent';
 
-const CreateNftCollectionScreen = () => {
-  const {createNftCollection, sendTransaction} = useWallet();
+const MintNftScreen = (props: any) => {
+  const {MintNft, ExecWrapperPromise, sendTransaction, parsedAddresses} =
+    useWallet();
   const {readPassword} = useSecurity();
 
-  const {goBack} = useNavigation<NavigationProp<RootStackParamList>>();
+  const {navigate} = useNavigation<NavigationProp<RootStackParamList>>();
   const {collapse} = useBottomSheet();
   const {openModal, closeModal} = useModal();
   const bottomSheet = useBottomSheet();
+
+  const nftCollection = props.route.params.from;
 
   const [error, setError] = useState('');
   const [loading, setLoading] = useState<string | undefined>(undefined);
@@ -42,14 +44,52 @@ const CreateNftCollectionScreen = () => {
     closeModal();
   }, [loading]);
 
-  const [collection, setCollection] = useState<CollectionOption>({
+  const [collection, setCollection] = useState<NftItemOption>({
     name: '',
-    description: '',
-    amount: 1,
+    resource: '',
   });
 
+  const [assignedId, setAssignedId] = useState<number>(-1);
+  const [destination, setDestination] = useState<string>('');
+  const [fullyMinted, setFullyMinted] = useState<boolean>(false);
+
+  useEffect(() => {
+    setDestination(
+      parsedAddresses.filter(
+        el => el.type_id == Destination_Types_Enum.PrivateWallet,
+      )[0]?.address,
+    );
+
+    ExecWrapperPromise(
+      'wallet.GetNftInfo',
+      [nftCollection.tokenId, -1].map(el => JSON.stringify(el)),
+    ).then(nftInfo => {
+      if (nftInfo.length == 0) {
+        setAssignedId(0);
+        return;
+      }
+      let tempAssignedId = -1;
+      let found = false;
+      for (let item of nftInfo) {
+        if (item.id > tempAssignedId + 1) {
+          found = true;
+          tempAssignedId += 1;
+          break;
+        }
+        tempAssignedId = item.id;
+      }
+      if (!found && tempAssignedId + 1 < nftCollection.supply) {
+        found = true;
+        tempAssignedId += 1;
+      } else if (!found) {
+        setFullyMinted(true);
+      }
+      if (found) setAssignedId(tempAssignedId);
+    });
+  }, []);
+
   const setCollectionProperty = (
-    type: keyof CollectionOption,
+    type: keyof NftItemOption,
     value: string | number,
   ) => {
     const temp = {...collection};
@@ -57,26 +97,45 @@ const CreateNftCollectionScreen = () => {
     setCollection(temp);
   };
 
-  const createCollection = async () => {
-    if (!collection?.name || !collection?.description) {
-      setError('Please fill the collection details.');
+  const mintNftItem = async () => {
+    if (!collection?.name || !collection?.resource) {
+      setError('Please fill the item details.');
       return;
+    }
+
+    if (
+      ['png', 'jpg', 'jpeg', 'gif'].indexOf(
+        collection?.resource.split('.').pop(),
+      ) == -1
+    ) {
+      setError('Supported formats are: png, jpg, jpeg and gif');
+      return;
+    }
+
+    if (assignedId == -1) {
+      setError('Please wait for ID assignment');
     }
 
     readPassword()
       .then(async spendingPassword => {
         setLoading('Creating transaction...');
-        createNftCollection(
-          collection['name'],
-          JSON.stringify({...collection, amount: undefined, name: undefined}),
-          collection['amount'],
+        MintNft(
+          nftCollection.tokenId,
+          assignedId,
+          destination,
+          JSON.stringify({
+            ...collection,
+            image: collection.resource,
+            resource: undefined,
+            attributes: {thumbnail_url: collection.resource},
+          }),
           spendingPassword,
         )
           .then(tx => {
             setLoading(undefined);
             bottomSheet.expand(
               <BottomSheetView>
-                <TopNavigation title="Confirm collection creation" />
+                <TopNavigation title="Confirm NFT mint" />
                 <Layout level="2" style={styles.card}>
                   <View style={styles.row}>
                     <Text category="headline" style={{marginRight: 16}}>
@@ -91,23 +150,12 @@ const CreateNftCollectionScreen = () => {
 
                   <View style={styles.row}>
                     <Text category="headline" style={{marginRight: 16}}>
-                      Description:
+                      Image URL:
                     </Text>
                     <Text
                       category="headline"
                       style={{flex: 1, flexWrap: 'wrap'}}>
-                      {collection['description']}
-                    </Text>
-                  </View>
-
-                  <View style={styles.row}>
-                    <Text category="headline" style={{marginRight: 16}}>
-                      Number of items:
-                    </Text>
-                    <Text
-                      category="headline"
-                      style={{flex: 1, flexWrap: 'wrap'}}>
-                      {collection['amount']}
+                      {collection['resource']}
                     </Text>
                   </View>
                 </Layout>
@@ -144,7 +192,7 @@ const CreateNftCollectionScreen = () => {
                       } else {
                         setLoading(undefined);
                         collapse();
-                        goBack();
+                        navigate('MainWalletScreen');
                       }
                     });
                   }}
@@ -187,7 +235,7 @@ const CreateNftCollectionScreen = () => {
   return (
     <Container useSafeArea>
       <KeyboardAwareScrollView>
-        <TopNavigationComponent title={'Create a private NFT collection'} />
+        <TopNavigationComponent title={'Mint a NFT'} />
 
         <Layout level="2" style={styles.inputCard}>
           <View style={styles.inputGroup}>
@@ -205,33 +253,26 @@ const CreateNftCollectionScreen = () => {
           </View>
           <View style={styles.inputGroup}>
             <Text category="headline" style={[styles.inputTitle]}>
-              Description:
+              Image URL:
             </Text>
             <Input
               style={[styles.inputField]}
-              value={collection['description']}
+              value={collection['resource']}
               onChangeText={value => {
-                setCollectionProperty('description', value);
+                setCollectionProperty('resource', value);
               }}
             />
           </View>
-          <View style={styles.inputGroup}>
-            <Text category="headline" style={[styles.inputTitle]}>
-              Number of items:
-            </Text>
-            <Input
-              style={[styles.inputField]}
-              value={collection['amount']?.toString()}
-              keyboardType={'number-pad'}
-              returnKeyType={'done'}
-              placeholder={'0'}
-              onChangeText={value => {
-                setCollectionProperty('amount', parseInt(value || '0'));
-              }}
-            />
-          </View>
-          <Button status={'primary-whisper'} onPress={() => createCollection()}>
-            Create
+          <Button
+            status={'primary-whisper'}
+            onPress={() => {
+              if (assignedId != -1) mintNftItem();
+            }}>
+            {fullyMinted
+              ? 'The collection is already fully minted'
+              : assignedId == -1
+              ? 'Assigning item ID...'
+              : 'Create item #' + assignedId}
           </Button>
           {error ? (
             <Text style={[styles.errorText]} center>
@@ -245,7 +286,7 @@ const CreateNftCollectionScreen = () => {
     </Container>
   );
 };
-export default CreateNftCollectionScreen;
+export default MintNftScreen;
 
 const styles = StyleSheet.create({
   inputCard: {
